refactor: lazy load about route
