--- src/app/common/default-image-alt.ts
@@ -1,1 +1,0 @@
-export const DEFAULT_IMAGE_ALT = 'Image'
