+++ src/app/common/images/default-alt.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_ALT = 'Image'
