+++ src/app/common/routing/get-canonical-url-for-path.ts
@@ -1,0 +1,8 @@
+import defaultMetadata from '../../../data/misc/metadata.json'
+
+export function getCanonicalUrlForPath(...pathSegments: ReadonlyArray<string>) {
+  return new URL(
+    pathSegments.join('/'),
+    new URL(defaultMetadata.canonicalUrl),
+  ).toString()
+}
