+++ src/app/common/routing/json-metadata.ts
@@ -1,0 +1,15 @@
+export interface JsonMetadata {
+  readonly name: string
+  readonly title: string
+  readonly description: string
+  readonly keywords?: string[] | undefined
+  readonly image?:
+    | {
+        readonly url: string
+        readonly alt: string
+        readonly width: number
+        readonly height: number
+        readonly mimeType: string
+      }
+    | undefined
+}
