+++ src/app/common/routing/navigator.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core'
 import { Router } from '@angular/router'
-import { NOT_FOUND_PATH } from '../routes'
 import { noop } from 'rxjs'
+import { NOT_FOUND_PATH } from './paths'
 
 @Injectable({
   providedIn: 'root',
