+++ src/app/common/routing/paths.ts
@@ -1,0 +1,7 @@
+import projectsPageMetadata from '../../../data/pages/projects.json'
+import notFoundPageMetadata from '../../../data/pages/404.json'
+import aboutPageMetadata from '../../../data/pages/about.json'
+
+export const PROJECTS_PATH = projectsPageMetadata.slug
+export const NOT_FOUND_PATH = notFoundPageMetadata.slug
+export const ABOUT_PATH = aboutPageMetadata.slug
