+++ src/app/common/routing/route-data.ts
@@ -1,0 +1,5 @@
+import { IPageSeoData } from '@ngaox/seo'
+
+export type RouteData = {
+  NgaoxSeo?: IPageSeoData
+}
