+++ src/app/projects/project-page/project-page.component.ts
@@ -2,16 +2,18 @@ import { Component, Input } from '@angular/core'
 import { ProjectsService } from '../projects.service'
 import { catchError, concatMap, map, Observable, of, tap } from 'rxjs'
 import { SeoService } from '@ngaox/seo'
-import { getCanonicalUrlForPath, getTitle, PROJECTS_PATH } from '../../routes'
-import { ImageResponsiveBreakpointsService } from '../../common/image-responsive-breakpoints.service'
-import { ImageResponsiveBreakpoints } from '../../common/image-responsive-breakpoints'
-import { NavigatorService } from '../../common/navigator.service'
+import { ImageResponsiveBreakpointsService } from '../../common/images/image-responsive-breakpoints.service'
+import { ImageResponsiveBreakpoints } from '../../common/images/image-responsive-breakpoints'
+import { NavigatorService } from '../../common/routing/navigator.service'
 import { AnyAssetsCollection } from './any-asset-collection'
 import { ProjectAssetsCollectionsService } from './project-assets-collections.service'
 import { SwiperOptions } from 'swiper/types'
 import { AssetsCollectionData } from './assets-collection-data'
 import { AssetsCollectionSize } from './assets-collection-size'
 import { AssetsCollectionType } from './assets-collection-type'
+import { getCanonicalUrlForPath } from '../../common/routing/get-canonical-url-for-path'
+import { PROJECTS_PATH } from '../../common/routing/paths'
+import { getTitle } from '../../common/routing/get-title'
 
 @Component({
   selector: 'app-project-page',
